+++ src/modules/admin/widgets/grids/traits/AssetColumnsTrait.php
@@ -7,8 +7,8 @@ namespace davidhirtz\yii2\media\modules\admin\widgets\grids\traits;
 use davidhirtz\yii2\media\assets\AdminAsset;
 use davidhirtz\yii2\media\models\interfaces\AssetInterface;
 use davidhirtz\yii2\media\models\interfaces\AssetParentInterface;
+use davidhirtz\yii2\skeleton\db\ActiveRecord;
 use davidhirtz\yii2\skeleton\helpers\Html;
-use davidhirtz\yii2\skeleton\modules\admin\widgets\grids\GridView;
 use davidhirtz\yii2\skeleton\widgets\fontawesome\Icon;
 use Yii;
 use yii\data\ActiveDataProvider;
@@ -17,12 +17,9 @@ use yii\db\ActiveRecordInterface;
 use yii\db\ExpressionInterface;
 use yii\helpers\Url;
 
-/**
- * @mixin GridView
- */
 trait AssetColumnsTrait
 {
-    public ?AssetParentInterface $parent = null;
+    public AssetParentInterface $parent;
 
     /**
      * @var int|ExpressionInterface|null the maximum number of assets loaded for `$parent`
@@ -54,33 +51,29 @@ trait AssetColumnsTrait
     }
 
     /**
-     * @param AssetInterface $model
+     * @param ActiveRecord&AssetInterface $model
      */
-    protected function getDeleteButton(ActiveRecordInterface $model): string
+    protected function getDeleteButton(ActiveRecordInterface $model, array $options = []): string
     {
-        $options = [
-            'class' => 'btn btn-danger btn-delete-asset d-none d-md-inline-block',
-            'data-confirm' => Yii::t('media', 'Are you sure you want to remove this asset?'),
-            'data-target' => '#' . $this->getRowId($model),
-            'data-ajax' => 'remove',
-        ];
-
+        $options['class'] ??= 'btn btn-danger btn-delete-asset d-none d-md-inline-block';
+        $options['data-confirm'] ??= Yii::t('media', 'Are you sure you want to remove this asset?');
 
         if (Yii::$app->getUser()->can('fileDelete', ['file' => $model->file])) {
             $options['data-delete-message'] = Yii::t('media', 'Permanently delete related files');
             $options['data-delete-url'] = Url::to(['file/delete', 'id' => $model->file_id]);
         }
 
-        return Html::a(Icon::tag('trash'), $this->getDeleteRoute($model), $options);
+        return parent::getDeleteButton($model, $options);
     }
 
-    protected function getFileUpdateButton(AssetInterface $asset): string
+    protected function getFileUpdateButton(AssetInterface $asset, array $options = []): string
     {
-        return Html::a(Icon::tag('image'), ['file/update', 'id' => $asset->file_id], [
+        return Html::a((string)Icon::tag('image'), ['file/update', 'id' => $asset->file_id], [
             'class' => 'btn btn-secondary d-none d-md-inline-block',
             'title' => Yii::t('media', 'Edit File'),
             'data-toggle' => 'tooltip',
             'target' => '_blank',
+            ...$options,
         ]);
     }